+++ airticketopen-20230117/alibabacloud_airticketopen20230117/models/_global_hotel_validate_price_shrink_request.py
@@ -19,13 +19,13 @@ class GlobalHotelValidatePriceShrinkRequest(DaraModel):
         # 
         # This parameter is required.
         self.account_no = account_no
-        # The number of adults per room.
+        # The total number of adults.
         # 
         # This parameter is required.
         self.adults = adults
-        # The number of children per room.
+        # The total number of children.
         self.children = children
-        # The list of children ages.
+        # The list of children\\"s ages.
         self.children_ages_shrink = children_ages_shrink
         # The offer key.
         #