+++ airticketopen-20230117/alibabacloud_airticketopen20230117/models/_search_response_body.py
@@ -553,6 +553,7 @@ class SearchResponseBodyDataSolutionListJourneyListSegmentList(DaraModel):
         departure_terminal: str = None,
         departure_time: str = None,
         equip_type: str = None,
+        fare_basis: str = None,
         flight_duration: int = None,
         marketing_airline: str = None,
         marketing_flight_no: str = None,
@@ -589,6 +590,7 @@ class SearchResponseBodyDataSolutionListJourneyListSegmentList(DaraModel):
         self.departure_time = departure_time
         # The aircraft type.
         self.equip_type = equip_type
+        self.fare_basis = fare_basis
         # The flight duration. Unit: minutes.
         self.flight_duration = flight_duration
         # The marketing airline (such as KA).
@@ -654,6 +656,9 @@ class SearchResponseBodyDataSolutionListJourneyListSegmentList(DaraModel):
 
         if self.equip_type is not None:
             result['equip_type'] = self.equip_type
+
+        if self.fare_basis is not None:
+            result['fare_basis'] = self.fare_basis
 
         if self.flight_duration is not None:
             result['flight_duration'] = self.flight_duration
@@ -724,6 +729,9 @@ class SearchResponseBodyDataSolutionListJourneyListSegmentList(DaraModel):
 
         if m.get('equip_type') is not None:
             self.equip_type = m.get('equip_type')
+
+        if m.get('fare_basis') is not None:
+            self.fare_basis = m.get('fare_basis')
 
         if m.get('flight_duration') is not None:
             self.flight_duration = m.get('flight_duration')